refactor: decompose digital-twin.js god file into 10 focused modules
Split 2,970-line services/digital-twin.js into:
- digital-twin-constants.js  — ENRICHMENT_CATEGORIES, SCALE_QUESTIONS
- digital-twin-helpers.js    — generateId, now, ensureSoulDir, callProviderAI
- digital-twin-meta.js       — cache, events, loadMeta/saveMeta/updateMeta/updateSettings
- digital-twin-documents.js  — CRUD for digital twin markdown files
- digital-twin-testing.js    — parseTestSuite, runTests, getTestHistory
- digital-twin-enrichment.js — questionnaire system, scale questions, list enrichment
- digital-twin-export.js     — 5 export format handlers + legacy portrait
- digital-twin-context.js    — CoS prompt injection (getDigitalTwinForPrompt)
- digital-twin-analysis.js   — traits, confidence, validation, contradiction detection
- digital-twin-import.js     — Goodreads/Spotify/Letterboxd/iCal imports + assessment analyzer
- digital-twin-status.js     — getDigitalTwinStatus health score aggregation

digital-twin.js is now a thin barrel re-export. All 88 tests pass.
